Started the Restructuring of the Development Project file
List of new Files with the functions or classes moved in are:
testing_operations
- StMTestImagePrep
- StMTestConnectionOperator
- DoImg
- StMTestGetFileFromDbFromUserId
- StMTestDeleteFileFromDbFromUserId
- StMTestSaveFileToDb
- outline_image

ui_operations
- Reset_Input_Images
- PlaceImageIn3D
