+++ __init__.py
@@ -7,162 +7,14 @@ bl_info = {
     "description": "The Inital UI skeleton",
     "category": "Development",
 }
-    # 3D Viewport area (find list of values here https://docs.blender.org/api/current/bpy_types_enum_items/space_type_items.html#rna-enum-space-type-items)
-    # Sidebar region (find list of values here https://docs.blender.org/api/current/bpy_types_enum_items/region_type_items.html#rna-enum-region-type-items)
-    # bpy.ops.wm.save_as_mainfile(filepath="c:\Users\James Burns\Documents\TestFile.blend")
     
 import bpy
-import cv2
 import numpy as np
-import os
 
-from dataclasses import dataclass
-from .db_operations import test_connection, save_file_to_db, get_files_by_user_id, delete_files_by_object_id # the . is on purpose. do not remove
-from .image_processing import prepare_image, test_feature_detection # the . is on purpose. do not remove
+from .testing_operations import StMTestImagePrep, StMTestConnectionOperator, DoImg, StMTestGetFileFromDbFromUserId, StMTestDeleteFileFromDbFromUserId, StMTestSaveFileToDb
+from .ui_operations import PlaceImageIn3D, Reset_Input_Images
 from .bcrypt_password import hash_password
 from .authentication import login_account, register_account
-import bcrypt # unsure
-
-#FutureReference
-#this will contain the filepath and the image plane name
-#this has not been implemented yet
-@dataclass
-class FilePathStructure:
-    filePath: str
-    ImagePlaneName: str
-
-GlobalFileImageStructArray = [] #this will eventually replace the two array under this
-globalPlaneArray = [] # a list of the names of the planes in Blender
-globalfilePathsArray = [] # a list of the file path we wan to keep track of
-UserSignedIn = False
-
-
-
-    
-class StMTestImagePrep(bpy.types.Operator):
-    bl_idname = "wm.prepare_image_operator"
-    bl_label = "Test Image Prep"
-
-    def execute(self, context):
-        test_feature_detection()
-        
-        #success = prepare_image(path)
-        #if success:
-        #    self.report({'INFO'}, "Image Prep Succesful!")
-        #else:
-        #    self.report({'ERROR'}, "Failed to Image Prep.")
-
-
-# wth is this? why is this here and not in a proper file? why does this method even exists? sigh.
-def outline_image(image_path, Extension, ImgName, Filedirectory):
-    """Read an image from a path, outline it, calculate the center of mass for the outlines, and draw a blue dot there."""
-    image = cv2.imread(image_path)
-
-    if image is None:
-        print("Error: Image not found or unable to load.")
-        return
-    
-    # temporary file size. adjusting files to the same scale can be beneficial for feature detectors
-    resized_image = cv2.resize(image, (800, 600))  
-    # grayscale reduces computational load
-    gray_image = cv2.cvtColor(resized_image, cv2.COLOR_BGR2GRAY) 
-    # noise reduction
-    blurred_image = cv2.GaussianBlur(gray_image, (5, 5), 0)
-    # canny edge detection emphasizes edges in the image
-    # we will most likely be using one of the two as feature detectors: ORB, AKAZE.
-    # both feature detection algorithms have positive results from this as they often rely on edge information to find key points.
-    edges = cv2.Canny(blurred_image, 50, 15, apertureSize=3)
-    #gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
-    #edges = cv2.Canny(gray, 50, 150, apertureSize=3)
-    contours, _ = cv2.findContours(edges, cv2.RETR_TREE, cv2.CHAIN_APPROX_SIMPLE)
-
-    # Check if any contours were found
-    if contours:
-        # Draw contours on the original image
-        cv2.drawContours(resized_image, contours, -1, (0, 255, 0), 2)
-        # Calculate the combined center of mass for all contours
-        totalX, totalY, totalArea = 0, 0, 0
-
-        for contour in contours:
-            M = cv2.moments(contour)
-
-            if M["m00"] != 0:
-                totalX += int(M["m10"])
-                totalY += int(M["m01"])
-                totalArea += M["m00"]
-
-        if totalArea != 0:
-            cX = int(totalX / totalArea)
-            cY = int(totalY / totalArea)
-            # Draw a blue dot at the combined center of mass
-            cv2.circle(resized_image, (cX, cY), 5, (255, 0, 0), -1)
-        else:
-            print("Error: Combined center of mass could not be calculated.")
-    else:
-        print("Error: No contours found.")
-
-    os.chdir(Filedirectory) #changes the directory to the folder where we are going to save the file
-    cv2.imwrite(ImgName + Extension, resized_image) #saves the image
-    os.chdir("..\\") #goes back one directory
-
-
-
-class PlaceImageIn3D(bpy.types.Operator):
-    bl_idname = "object.place_image_in_space"
-    bl_label ="Place Images"
-
-    PlaneArray = globalPlaneArray
-    filePathsArray = globalfilePathsArray
-
-    def execute(self, context):
-        #this will keep count of the views were have captured
-        Itervalue = 0
-        #this will be a folder in the sketch-to-Mesh project. This will hold the Image processed
-        ImageDiretoryForNewImage = "ImageFolder"
-        #this will eventually need to move to somewhere more accessable
-        #this is only here for now
-        filePathsArray = [bpy.context.scene.front_views_file_path, bpy.context.scene.back_views_file_path,  bpy.context.scene.side_views_file_path ]
-    
-        for file_path in filePathsArray : 
-            if file_path :
-                #this is used for the new image. We want to save the new image as the same type of file as the first
-                Extension =  file_path[file_path.rfind("."): ] 
-
-                #this is the file name for the image we are creating
-                ImgName = "View" + str(Itervalue)
-                self.PlaneArray.insert(Itervalue, ImgName) # allows us to access the plane after creation
-
-                #this is where we want to save the picture so i can be reaccessed
-                outline_image(file_path, Extension, ImgName, ImageDiretoryForNewImage)
-
-                #this creates a new file path to the image we just saved
-                NewFilePath = os.path.abspath(ImageDiretoryForNewImage + "\\" + ImgName + Extension) 
-                
-                if NewFilePath:
-                    filename = os.path.basename(NewFilePath)
-                    FileDirectory = NewFilePath[: NewFilePath.rfind("\\")] + "\\"
-
-                    #bpy.ops.import_image.to_plane(files=[{"name":filename, "name":filename}], directory=FileDirectory, relative=False)
-                    bpy.ops.import_image.to_plane(files=[{"name":filename, "name":filename}], directory=FileDirectory, relative=False)
-                    #we set the rotation and location of each plane
-                    bpy.data.objects[ImgName].select_set(True)
-                    match Itervalue :
-                        case 1: bpy.ops.transform.translate(value=(-0.01, 0 , 0), orient_type='GLOBAL', orient_matrix=((1, 0, 0), (0, 1, 0), (0, 0, 1)), orient_matrix_type='GLOBAL', constraint_axis=(False, True, False), mirror=False, use_proportional_edit=False, proportional_edit_falloff='SMOOTH', proportional_size=1, use_proportional_connected=False, use_proportional_projected=False, snap=False, snap_elements={'INCREMENT'}, use_snap_project=False, snap_target='CLOSEST', use_snap_self=True, use_snap_edit=True, use_snap_nonedit=True, use_snap_selectable=False, alt_navigation=True)
-                        case 2: bpy.context.object.rotation_euler[2] = 0
-                else:
-                    match Itervalue:
-                        case 0: MissingView = "FontView"
-                        case 1: MissingView = "BackView"
-                        case 2: MissingView = "SideView"
-                    self.report({'ERROR'}, "No inputted Image for" + MissingView)
-                Itervalue = Itervalue + 1
-            else:
-                self.report({'ERROR'}, "No inputted Image.")
-                Itervalue = Itervalue + 1
-
-        return {'FINISHED'}
-
-
    
 # this contains the main layout for the Sketch to mesh program
 # to link up functions with the buttons
@@ -181,45 +33,6 @@ class VIEW3D_PT_Sketch_To_Mesh_Panel(bpy.types.Panel):
     def draw(self, context): 
         layout = self.layout
 
-# class that executes test_connection from db_operations
-# will be deleted in beta versions
-class StMTestConnectionOperator(bpy.types.Operator):
-    bl_idname = "wm.test_connection_operator"
-    bl_label = "Test Database Connection"
-
-    def execute(self, context): 
-        success = test_connection()
-        if success:
-            self.report({'INFO'}, "Connection to MongoDB successful!")
-        else:
-            self.report({'ERROR'}, "Failed to connect to MongoDB.")
-        return {'FINISHED'}
-
-class Reset_Input_Images(bpy.types.Operator): 
-    bl_idname = "object.reset_selected_images"
-    bl_label = "Reset_Images"
-    
-    myFilePathArray = globalfilePathsArray # saves the entries in the global list to be deleted
-    ImagePanelArray = globalPlaneArray
-    globalPlaneArray.clear()
-    globalfilePathsArray.clear() # clears the global list 
-
-    def execute(self, context):
-        Itervalue = 0
-
-        for filePath in self.myFilePathArray :
-            #Finds the Images Saved
-            ImageFilePath = os.path.abspath("ImageFolder\\" + "View" + str(Itervalue) + filePath[filePath.rfind("."): ] ) 
-            # if we find that file we will delete it
-            if ImageFilePath:
-                os.remove(ImageFilePath)
-
-        for images in self.ImagePanelArray:
-            # selects the image plane in the array 
-            bpy.data.objects[images].select_set(True)
-            #deletes the image plane in the array
-            bpy.ops.object.delete(use_global=False, confirm=False)
-        return {'FINISHED'}
 
 class VIEW3D_PT_Sketch_To_Mesh_Views_Panel(bpy.types.Panel):  
     bl_label = "View"
@@ -238,27 +51,6 @@ class VIEW3D_PT_Sketch_To_Mesh_Views_Panel(bpy.types.Panel):
         row.prop(context.scene, "side_views_file_path", text="Side View")
         row = layout.row()
         row.operator("object.reset_selected_images", text="Reset Images")
-
-
-
-class DoImg(bpy.types.Operator):
-    bl_idname = "object.do_img"
-    bl_label = "Place Image"
-
-    #Property that holds the filepath that will be used to insert the image
-    myFilePath = ""
-    #bpy.props.StringProperty(subtype="FILE_PATH")
-
-    #This is the function that inserts the image into blender
-    def execute(self, context):
-        bpy.ops.object.load_reference_image(filepath=self.myFilePath)
-        return {'FINISHED'}
-    #This is a function that opens a file explorer
-    #THIS DOES NOT DO ANYTHING I just think its going to be useful to have later
-    def invoke(self, context, event):
-        # Open a file browser to select a file
-        context.window_manager.fileselect_add(self)
-        return {'RUNNING_MODAL'}
 
 
 class VIEW3D_PT_Sketch_To_Mesh_Align_Views_Panel(bpy.types.Panel):  
@@ -300,28 +92,6 @@ class VIEW3D_PT_Sketch_To_Mesh_MeshSettings_Panel(bpy.types.Panel):
         row.operator("mesh.primitive_cube_add", text="Export Mesh")
 
 
-###this is a example class ###
-class ExampleOperator(bpy.types.Operator):
-    bl_idname = "load.ExampleName" #the first word is the type of thing your doing(probally should look this up) follow by '.' and the name you want to use to call the operator
-    bl_label = "Load Image"
-    bl_options = {'REGISTER', 'UNDO'}
-
-    #properities go here
-
-    #this is where you define what you want to happen
-    def execute(self, context):
-      # Do Something
-        return {'FINISHED'}
-    
-    #this is how you would call the operators in other classes
-    #layout.operator("load.ExampleName", text="")
-    #text is what you want displayed
-
-    #there need to be called in the register and unregister definetions respectively close to the bottom of the script
-    #bpy.utils.register_class(LoadImageOperator) 
-    #bpy.utils.unregister_class(LoadImageOperator)
-    ###this is the end of the example class ###
-    
 class DataBaseLogin(bpy.types.Operator):
     bl_idname = "wm.database_login_popup"
     bl_label = "Database Register/Login"
@@ -371,49 +141,6 @@ class DataBaseLogin(bpy.types.Operator):
     def invoke(self, context, event):
         return context.window_manager.invoke_props_dialog(self)
 
-class StMTestSaveFileToDb(bpy.types.Operator):
-    bl_idname = "wm.save_file_to_db_operator"
-    bl_label = "Test Saving File"
-
-    def execute(self, context):
-        
-        save_file_to_db("123") # needs a file path but are not using
-
-        return {'FINISHED'}
-    
-
-class StMTestGetFileFromDbFromUserId(bpy.types.Operator):
-    bl_idname = "wm.get_file_from_db_operator"
-    bl_label = "Test Getting File"
-
-    def execute(self, context):
-        
-        result = get_files_by_user_id("123") # 123 since the only document in the db is 123
-        
-        for document in result:
-            # removed the bin data because it was too annoying as the output. fileEncoded: {document['fileEncoded']}
-            print(f"objectId: {document['_id']}, filename: {document['fileName']}, userId: {document['userId']}, insertedDate: {document['insertedDate']}")
-
-        return {'FINISHED'}
-    
-class StMTestDeleteFileFromDbFromUserId(bpy.types.Operator):
-    bl_idname = "wm.delete_file_from_db_operator"
-    bl_label = "Test Deleting File"
-
-    def execute(self, context):
-        
-        objectId = "65ccec75d26b1d7703fb3a0a"
-        result = delete_files_by_object_id(objectId) # 123 since the only document in the db is 123
-        
-        if (result == 0):
-            print("No files deleted. Check ObjectID")
-
-        if (result == 1):
-            print(f"objectId: {objectId} file successfully deleted.")
-            self.report({'INFO'}, "File successfully deleted.")
-        
-
-        return {'FINISHED'}
 
 class VIEW3D_PT_Sketch_To_Mesh_Testing(bpy.types.Panel):  
     bl_label = "Testing"
@@ -424,26 +151,18 @@ class VIEW3D_PT_Sketch_To_Mesh_Testing(bpy.types.Panel):
 
     def draw(self, context):
         layout = self.layout
-
         row = layout.row()
         row.operator("wm.test_connection_operator", text="Test Connection")
-
         row = layout.row()
         row.operator("wm.database_login_popup", text="Access Database")
-        
         row = layout.row()
         row.operator("wm.prepare_image_operator", text="Test Image Prep")
-        
         row = layout.row()
         row.operator("wm.save_file_to_db_operator", text="Save File to DB")
-
         row = layout.row()
         row.operator("wm.get_file_from_db_operator", text="Get File from DB")
-
         row = layout.row()
         row.operator("wm.delete_file_from_db_operator", text="Delete File from DB")
-
-
 
 
 def register():
