+++ image_processing.py
@@ -2,8 +2,6 @@ import cv2
 import numpy as np
 import os
 
-# this will be called once the images are ready
-
 def prepare_image(image_path):
     
     image = cv2.imread(image_path)
@@ -40,7 +38,6 @@ def prepare_image(image_path):
     except Exception as e:
         print(f"Error: {e}")
         return False
-    
 
 def match_features(descriptors1, descriptors2, method='ORB'):
     # using ORB and AKAZE for testing
@@ -65,7 +62,6 @@ def draw_matches(image1, keypoints1, image2, keypoints2, matches):
 
     matched_image = cv2.drawMatches(image1, keypoints1, image2, keypoints2, matches, None, flags=cv2.DrawMatchesFlags_NOT_DRAW_SINGLE_POINTS)
     return matched_image
-
 
 def detect_and_describe_akaze(image_path):
     image = cv2.imread(image_path)
